chore: make changes to real file, delete test file
